WLT-297: moved hash and sign interfaces
