+++ platformpolicy/commoncrypto/hash/wrapper.go
