+++ platformpolicy/commoncrypto/wrapper_test.go
@@ -14,24 +14,24 @@
 // limitations under the License.
 //
 
-package internal
+package commoncrypto
 
 import (
-	"github.com/insolar/insolar/platformpolicy/commoncrypto"
 	"testing"
 
-	"github.com/insolar/insolar/component"
-	"github.com/insolar/insolar/platformpolicy/internal/hash"
-	"github.com/insolar/insolar/platformpolicy/internal/sign"
-	"github.com/insolar/insolar/testutils"
 	"github.com/stretchr/testify/assert"
+
+	"github.com/insolar/insolar/component"
+	"github.com/insolar/insolar/platformpolicy/commoncrypto/hash"
+	"github.com/insolar/insolar/platformpolicy/commoncrypto/sign"
+	"github.com/insolar/insolar/testutils"
 )
 
 func TestEcdsaMarshalUnmarshal(t *testing.T) {
 	count := 10000
 	data := testutils.RandomRef()
 
-	kp := commoncrypto.NewKeyProcessor()
+	kp := NewKeyProcessor()
 	provider := sign.NewECDSAProvider()
 
 	cm := component.Manager{}
